+++ Sources/OpenVPN/Cross/Internal/Errors/TLSError.swift
@@ -3,6 +3,7 @@
 // SPDX-License-Identifier: GPL-3.0
 
 internal import _PartoutOpenVPN_C
+internal import _PartoutVendorsTLSCore_C
 
 enum TLSError: Error {
     case missingCA
