+++ Sources/Vendors/Crypto/CryptoCore_C/dummy.c
@@ -1,0 +1,5 @@
+/*
+ * SPDX-FileCopyrightText: 2025 Davide De Rosa
+ *
+ * SPDX-License-Identifier: GPL-3.0
+ */
