+++ Sources/Vendors/OpenSSL/include/shim.h
